# Conflicts:
#	README.rst
#	setup.py
#	umap/sparse.py
#	umap/umap_.py
#	umap/utils.py

+++ umap/utils.py
@@ -126,7 +126,7 @@ def make_heap(n_points, size):
     -------
     heap: An ndarray suitable for passing to other numba enabled heap functions.
     """
-    result = np.zeros((3, n_points, size))
+    result = np.zeros((3, int(n_points), int(size)), dtype=np.float64)
     result[0] = -1
     result[1] = np.infty
     result[2] = 0
